move Assitant CRUD APIs to api folder
