+++ core/account/serializers.py
@@ -8,10 +8,7 @@ class ProfessorSerializers(serializers.ModelSerializer):
         model = Professor
         fields = '__all__'
 
-class AssistantSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Assistant
-        fields = '__all__'
+
 
 class TermSerializer(serializers.ModelSerializer):
     class Meta:
