+++ core/account/views.py
@@ -5,7 +5,7 @@ from rest_framework import status, serializers
 from rest_framework.response import Response
 from .models import Assistant, Professor
 from lesson.models import Term
-from .serializers import AssistantSerializer, ProfessorSerializers, TermSerializer
+from .serializers import ProfessorSerializers, TermSerializer
 
 
 # Create your views here.
@@ -17,53 +17,6 @@ class RegisterProfessorView(ListCreateAPIView):
     permission_classes = [IsAdminUser]
 
 
-class CreateAssistantAPIView(ListCreateAPIView):
-    """
-    Make the user as an Assistant
-    """
-    serializer_class = AssistantSerializer
-    permission_classes = [IsAdminUser]
-    queryset = Assistant.objects.all()
-
-    def perform_create(self, serializer):
-        data = serializer.validated_data
-        existing_assistant = Assistant.objects.filter(user_id=data['user'])
-        if existing_assistant.exists():
-            raise serializers.ValidationError("An assistant with the same attributes already exists.")
-        serializer.save()
-
-    def create(self, request, *args, **kwargs):
-        try:
-            return super().create(request, *args, **kwargs)
-        except serializers.ValidationError as e:
-            return Response({'error': str(e)}, status=status.HTTP_409_CONFLICT)
-
-        except Exception as e:
-            return super().handle_exception(e)
-
-class AssistantAPIView(RetrieveUpdateDestroyAPIView):
-    """
-    Retrieve, update, delete an Assistant
-    """
-    serializer_class = AssistantSerializer
-    permission_classes = [IsAdminUser]
-    queryset = Assistant.objects.all()
-
-    def perform_update(self, serializer):
-        data = serializer.validated_data
-        existing_assistant = Assistant.objects.filter(user_id=data['user'])
-        if existing_assistant.exists():
-            raise serializers.ValidationError("An assistant with the same attributes already exists.")
-        serializer.save()
-
-    def update(self, request, *args, **kwargs):
-        try:
-            return super().update(request, *args, **kwargs)
-        except serializers.ValidationError as e:
-            return Response({'error': str(e)}, status=status.HTTP_409_CONFLICT)
-
-        except Exception as e:
-            return super().handle_exception(e)
 
 class CreateTermAPIView(ListCreateAPIView):
     """
