+++ core/api/v1/serializers/account_serilizers.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+from account.models import Assistant
+
+
+class AssistantSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Assistant
+        fields = '__all__'
