+++ core/core/urls.py
@@ -28,5 +28,5 @@ urlpatterns = [
     path('account/', include('account.urls')),
     path('api/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
     path('api/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
-    path('api/v1', include('api.vi.urls'), name='api-url'),
+    path('api/v1/', include('api.v1.urls'), name='api-url'),
 ]
